update for ConceptionKind Management features

+++ src/main/java/com/viewfunction/docg/element/visualizationComponent/payload/common/RelationshipEdgeData.java
@@ -6,7 +6,7 @@ import elemental.json.JsonObject;
 
 public class RelationshipEdgeData implements JsonSerializable {
 
-    private int weight;
+    private long weight;
     private String name;
     private String desc;
     private String id;
@@ -15,7 +15,7 @@ public class RelationshipEdgeData implements JsonSerializable {
 
     public RelationshipEdgeData(){}
 
-    public RelationshipEdgeData(String name,String desc,String id,String source,String target,int weight){
+    public RelationshipEdgeData(String name,String desc,String id,String source,String target,long weight){
         setName(name);
         setDesc(desc);
         setId(id);
@@ -24,21 +24,21 @@ public class RelationshipEdgeData implements JsonSerializable {
         setWeight(weight);
     }
 
-    public int increaseWeight(){
+    public long increaseWeight(){
         setWeight(getWeight() + 1);
         return getWeight();
     }
 
-    public int decreaseWeight(){
+    public long decreaseWeight(){
         setWeight(getWeight() - 1);
         return getWeight();
     }
 
-    public int getWeight() {
+    public long getWeight() {
         return weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(long weight) {
         this.weight = weight;
     }
